requirements.txt
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.1.0
pylint==2.6.0
pyTelegramBotAPI==3.7.6
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.3
waitress==1.4.4
Werkzeug==1.0.1
wrapt==1.12.1